Feature/rev buy crypto flow refactor

+++ domain/src/main/java/com/multimoney/domain/repository/CryptoRepository.kt
@@ -73,7 +73,7 @@ interface CryptoRepository {
         market: String,
         commissionAmount: Double,
         taxAmount: Double,
-        accountToken: Double,
+        accountToken: Long,
         exchangeRate: Double,
         idBrand: Int,
         user: String,
